highscore-cloudrun-service/requirements.txt
functions-framework==3.*
google-cloud-storage==3.*
